feat): Implement robust hook failure detection and auto-reload
Reworks the script's core initialization logic to prevent a race condition where the game's WebSocket connects before the script can intercept the necessary codecs.

If such failure is detected, the script now displays a user-friendly "Re-attempting Connection..." message and automatically reloads the page to ensure a successful injection. With the option to not reload and continue playing, without the mod enabled. This significantly improves the mod's reliability and user experience, especially on slower connections or network conditions.

Also, moved Packet Docs and added the original game code to the utlity mod folder.